requirements.txt
Flask==3.1.1
Flask-Mail==0.10.0
stripe==12.3.0
python-dotenv==1.1.1
gunicorn==23.0.0
